Add connection manager and update project properties
This commit introduces a new connection manager for the Phoenix database in `VENJJCR20.Phoenix.conmgr`, along with several parameters for connection configuration. The `IS-Migracion-Polaris.dtproj` file has been updated to reflect these changes, including the addition of new parameters and the removal of obsolete ones.

Additionally, the version builds for `PqTransformaDMS-Dra01Takeover` and `PqCrearDMS-Dea01Takeover` have been incremented, and their GUIDs have been updated to reflect the latest changes. Enhancements in the `PqCrearDMS.dtsx` file include the replacement of the executable with a new one that improves data processing capabilities, along with the addition of new input and output columns, derived column expressions, and graphical layout adjustments.